Rename toggles type file
