+++ src/components/Checkbox/MUI/Checkbox.tsx
@@ -1,6 +1,6 @@
 import { config } from '@root/app.config'
 import { FormControlLabel, Checkbox as MuiCheckbox, CheckboxProps as MuiProps } from '@mui/material'
-import { CheckboxProps } from '@apptypes/toggles'
+import { CheckboxProps } from '@apptypes/toggle'
 import { getStyles, toggleSetting } from '@utils/toggles/common'
 
 export const Checkbox = (props: CheckboxProps) => {
